refactor: created seperate save collection form component
